+++ crate_anon/anonymise/config.py
@@ -629,7 +629,9 @@ class Config(object):
         )
         self.denylist_filenames = cfg.opt_multiline(AK.DENYLIST_FILENAMES)
         self.denylist_files_as_phrase_lines = cfg.opt_bool(
-            AK.DENYLIST_FILES_AS_PHRASE_LINES)
+            AK.DENYLIST_FILES_AS_PHRASE_LINES, False)
+        self.denylist_phrases_flexible_whitespace = cfg.opt_bool(
+            AK.DENYLIST_PHRASES_FLEXIBLE_WHITESPACE, False)
         self.phrase_alternative_word_filenames = cfg.opt_multiline(
             AK.PHRASE_ALTERNATIVE_WORD_FILENAMES)
         self.scrub_all_numbers_of_n_digits = cfg.opt_multiline_int(
@@ -657,12 +659,13 @@ class Config(object):
         )
         self.denylist = WordList(
             filenames=self.denylist_filenames,
-            as_phrase_lines=self.denylist_files_as_phrase_lines,
+            as_phrases=self.denylist_files_as_phrase_lines,
             replacement_text=self.replace_nonspecific_info_with,
             hasher=self.change_detection_hasher,
             at_word_boundaries_only=(
                 self.anonymise_strings_at_word_boundaries_only),
             max_errors=0,
+            regex_method=self.denylist_phrases_flexible_whitespace,
         )
         self.nonspecific_scrubber = NonspecificScrubber(
             replacement_text=self.replace_nonspecific_info_with,