+++ crate_anon/anonymise/constants.py
@@ -226,6 +226,7 @@ class AnonymiseConfigKeys:
     ANONYMISE_STRINGS_AT_WORD_BOUNDARIES_ONLY = "anonymise_strings_at_word_boundaries_only"  # noqa
     DENYLIST_FILENAMES = "denylist_filenames"
     DENYLIST_FILES_AS_PHRASE_LINES = "denylist_files_as_phrase_lines"
+    DENYLIST_PHRASES_FLEXIBLE_WHITESPACE = "denylist_phrases_flexible_whitespace"  # noqa
     DEPRECATED_BLACKLIST_FILENAMES = "blacklist_filenames"
     DEPRECATED_WHITELIST_FILENAMES = "whitelist_filenames"
     MIN_STRING_LENGTH_FOR_ERRORS = "min_string_length_for_errors"
@@ -402,6 +403,7 @@ DEMO_CONFIG = rf"""# Configuration file for CRATE anonymiser (crate_anonymise).
 {_AK.ALLOWLIST_FILENAMES} =
 {_AK.DENYLIST_FILENAMES} =
 {_AK.DENYLIST_FILES_AS_PHRASE_LINES} = False
+{_AK.DENYLIST_PHRASES_FLEXIBLE_WHITESPACE} = False
 {_AK.PHRASE_ALTERNATIVE_WORD_FILENAMES} =
 {_AK.SCRUB_ALL_NUMBERS_OF_N_DIGITS} =
 {_AK.SCRUB_ALL_UK_POSTCODES} = False